Refactor initial state w/r/t hash
This simplifies and improves the process by which the Drawer is opened
by an existing hash. By setting `initState` instead of firing an action,
we avoid having a wierd number of actions on instantiation:

Consider the following situation:

- The drawer does not have a `data-state` attribute.
- A unique `initState` is described in the settings.
- A hash also applies when the drawer is instantiated.

This means the drawer will mutate twice, rapidly, on instatiation:

- From `undefied` to `initState`
- from `initState` to `hashState`

If any of these states are hidden, we could have even more mutations.
This would require implementations to consider this slightly ridiculous
chain of events, complicating things quite a bit.

This change should make it much more certain that the startup behavior
w/r/t to mutation events should be much more predictable (baring any
custom actions, which cannot reasonably be anticipated).