Rewrite more-or-less finished
It needs some additional testing, and some QoL and TS updates, but it
looks like everything works. Next step is to re-write the documentation
to reflect the new shape of the API. After that I'd like to take some
time and make sure I /like/ that API, since once we release I don't want
to change it again.